feat: Complete CLI modularization with enhanced functionality and bug fixes
This commit completes the CLI modularization refactoring and resolves critical
compatibility issues that were causing test failures.

CLI Modularization (Phases 1-3):
- Split monolithic CLI into specialized modules
- Maintained backward compatibility layer
- Added enhanced functionality with new flags
- Fixed missing imports and test compatibility issues

Key Features:
- --prt-debug-info flag for system diagnostics
- --classic flag for forced CLI mode
- Complete modular architecture with 26+ new files
- Zero test regressions, all functionality preserved

Validation: All CLI tests pass (8/8), debug info tests pass (13/13)
