+++ prt_src/cli_modules/bootstrap/__init__.py
@@ -1,0 +1,6 @@
+"""
+CLI Bootstrap Package
+
+Application initialization and setup functions.
+These modules handle database setup, health checks, and application launch.
+"""
