+++ prt_src/cli_modules/ui/__init__.py
@@ -1,0 +1,6 @@
+"""
+CLI UI Utilities Package
+
+Reusable display and prompt utilities shared across handlers.
+These utilities provide consistent user interaction patterns.
+"""
